setup map submodules
